moved test to treenode_test
